refactor: 예약/결제 기능 - 레이어드 아키텍처 → 클린 아키텍처 전환 (Infrastructure Layer 작성)
