--- src/main/java/kr/hhplus/be/server/reservation/PayHistoryRepository.java
@@ -1,7 +1,0 @@
-package kr.hhplus.be.server.reservation;
-
-import kr.hhplus.be.server.reservation.domain.PayHistory;
-import org.springframework.data.jpa.repository.JpaRepository;
-
-public interface PayHistoryRepository extends JpaRepository<PayHistory, Integer> {
-}
