+++ src/main/java/kr/hhplus/be/server/reservation/domain/PayHistory.java
@@ -5,6 +5,7 @@ import lombok.AllArgsConstructor;
 import lombok.Builder;
 import lombok.Getter;
 import lombok.NoArgsConstructor;
+import org.hibernate.annotations.CreationTimestamp;
 
 import java.time.LocalDateTime;
 import java.util.UUID;
@@ -39,7 +40,8 @@ public class PayHistory {
 
     private String reason;
 
-    @Column(name="created_at", insertable = false)
+    @CreationTimestamp
+    @Column(name="created_at", updatable = false)
     private LocalDateTime payAt;
 
 }
