+++ src/main/java/kr/hhplus/be/server/reservation/domain/PayHistoryRepository.java
@@ -1,0 +1,5 @@
+package kr.hhplus.be.server.reservation.domain;
+
+public interface PayHistoryRepository {
+    PayHistory save(PayHistory payHistory);
+}
