+++ src/main/java/kr/hhplus/be/server/reservation/infrastructure/persistence/jpa/JpaPayHistoryRepository.java
@@ -1,0 +1,9 @@
+package kr.hhplus.be.server.reservation.infrastructure.persistence.jpa;
+
+import kr.hhplus.be.server.reservation.domain.PayHistory;
+import org.springframework.data.jpa.repository.JpaRepository;
+
+public interface JpaPayHistoryRepository extends JpaRepository<PayHistory, Integer> {
+}
+
+
